+++ webextension/data/options-api.js
@@ -1,30 +1,67 @@
 'use strict';
 
-function getPreferences(prefId){
-	let defaultSettings = options_default;
-	if(typeof localStorage.getItem(prefId) != "undefined" && localStorage.getItem(prefId) != null){
-		let current_pref = localStorage.getItem(prefId);
+function getPreference(prefId){
+	let options = (chrome.extension.getBackgroundPage() != null)? chrome.extension.getBackgroundPage().optionsData.options : optionsData.options;
+	let defaultSettings = (chrome.extension.getBackgroundPage() != null)? chrome.extension.getBackgroundPage().optionsData.options_default : optionsData.options_default;
+	
+	let currentPreferences = (chrome.extension.getBackgroundPage() != null)? chrome.extension.getBackgroundPage().appGlobal.currentPreferences : appGlobal.currentPreferences;
+	if(currentPreferences.hasOwnProperty(prefId)){
+		let current_pref = currentPreferences[prefId];
 		switch(typeof defaultSettings[prefId]){
 			case "string":
 				return current_pref;
 				break;
 			case "number":
-				return parseInt(current_pref);
+				if(isNaN(parseInt(current_pref))){
+					console.warn(`${prefId} is not a number (${current_pref})`);
+					return defaultSettings[prefId];
+				} else if(typeof options[prefId].minValue == "number" && parseInt(current_pref) < options[prefId].minValue){
+					return options[prefId].minValue;
+				} else if(typeof options[prefId].maxValue == "number" && parseInt(current_pref) > options[prefId].maxValue){
+					return options[prefId].maxValue;
+				} else {
+					return parseInt(current_pref);
+				}
 				break;
 			case "boolean":
 				return getBooleanFromVar(current_pref);
+				break;
+			case "undefined":
+				console.warn(`The setting "${prefId}" has no default value`);
+				return current_pref;
+				break;
 			default:
-				return
+				console.warn(`Unknown default type for the setting ${prefId}: ${typeof defaultSettings[prefId]}`);
+				return current_pref;
 		}
-		return localStorage.getItem(prefId);
 	} else if(typeof defaultSettings[prefId] != "undefined"){
 		console.warn(`Preference ${prefId} not found, using default`);
 		savePreference(prefId, defaultSettings[prefId]);
 		return defaultSettings[prefId];
+	} else {
+		//console.warn(`Preference ${prefId} not found, no default`);
 	}
 }
 function savePreference(prefId, value){
-	localStorage.setItem(prefId, value);
+	let options = (chrome.extension.getBackgroundPage() != null)? chrome.extension.getBackgroundPage().optionsData.options : optionsData.options;
+	let defaultSettings = (chrome.extension.getBackgroundPage() != null)? chrome.extension.getBackgroundPage().optionsData.options_default : optionsData.options_default;
+	let currentPreferences = (chrome.extension.getBackgroundPage() != null)? chrome.extension.getBackgroundPage().appGlobal.currentPreferences : appGlobal.currentPreferences;
+	if(options.hasOwnProperty(prefId) && options[prefId].type == "integer"){
+		if(typeof options[prefId].minValue == "number" && parseInt(value) < options[prefId].minValue){
+			value = options[prefId].minValue;
+		} else if(typeof options[prefId].maxValue == "number" && parseInt(value) > options[prefId].maxValue){
+			value = options[prefId].maxValue;
+		}
+	}
+	if(typeof defaultSettings[prefId] == "boolean" || typeof defaultSettings[prefId] == "number"){
+		value = value.toString();
+	}
+	currentPreferences[prefId] = value;
+	chrome.storage.local.set({[prefId] : value}, function() {
+		if(typeof chrome.runtime.lastError == "object" && chrome.runtime.lastError != null){
+			console.dir(chrome.runtime.lastError);
+		}
+	});
 }
 function getBooleanFromVar(string){
 	switch(typeof string){