+++ webextension/data/options-data.js
@@ -1,6 +1,7 @@
 'use strict';
 
-var options = {
+var optionsData = {};
+optionsData.options = {
 	"custom_lstu_server": {
 		"title": "Custom LSTU server",
 		"description": "LSTU server, for example https://lstu.fr/",
@@ -9,16 +10,16 @@ var options = {
 	}
 }
 
-var options_default = {};
-var options_default_sync = {};
+optionsData.options_default = {};
+optionsData.options_default_sync = {};
 
-for(var id in options){
-	var option = options[id];
+for(var id in optionsData.options){
+	var option = optionsData.options[id];
 	if(typeof option.value != "undefined"){
-		options_default[id] = option.value;
+		optionsData.options_default[id] = option.value;
 		
 		if(!(typeof option.sync == "boolean" && option.sync == false)){
-			options_default_sync[id] = option.value;
+			optionsData.options_default_sync[id] = option.value;
 		}
 	}
 }